feat(NPC navigation with fences): Add new navigation region when the fences is calculted with the polygon
When we create a NavigationRegion2D we can add a group explicit and take the source geometry group to create a group to objects in node with the same name, with that, when we Bake navigation polygon in navigationRegion2d the polygon is modified to exclude the fences.

In the cow or chicken nagivationagent2d we need to set path postprocessing to edgecentered, better result.
And a less path desired distance.

BUUT, we have 2 bug (but don't be considerate for the moment in the tutorial, we will see later) :
- The geometry group don't work with tree object, the polygon bake is the same, even if i add a tree in the polygonregion2d
- The cow and chicken don't collide with the player

We have now finish the tutorial 12, we are now here :
https://youtu.be/it0lsREGdmc?list=PLWTXKdBN8RZe3ytf6qdR4g1JRy0j-93v9&t=11966